// web/src/api/generated/models/InvitationView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI ist der verbindliche Vertrag.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface InvitationView
 */
export interface InvitationView {
    /**
     * 
     * @type {Date}
     * @memberof InvitationView
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof InvitationView
     */
    expiresAt: Date;
    /**
     * 
     * @type {string}
     * @memberof InvitationView
     */
    id: string;
}

/**
 * Check if a given object implements the InvitationView interface.
 */
export function instanceOfInvitationView(value: object): value is InvitationView {
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('expiresAt' in value) || value['expiresAt'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    return true;
}

export function InvitationViewFromJSON(json: any): InvitationView {
    return InvitationViewFromJSONTyped(json, false);
}

export function InvitationViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): InvitationView {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': (new Date(json['createdAt'])),
        'expiresAt': (new Date(json['expiresAt'])),
        'id': json['id'],
    };
}

export function InvitationViewToJSON(json: any): InvitationView {
    return InvitationViewToJSONTyped(json, false);
}

export function InvitationViewToJSONTyped(value?: InvitationView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'createdAt': value['createdAt'].toISOString(),
        'expiresAt': value['expiresAt'].toISOString(),
        'id': value['id'],
    };
}