// web/src/api/generated/models/IssuedInvitationView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface IssuedInvitationView
 */
export interface IssuedInvitationView {
    /**
     * 
     * @type {Date}
     * @memberof IssuedInvitationView
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof IssuedInvitationView
     */
    expiresAt: Date;
    /**
     * 
     * @type {string}
     * @memberof IssuedInvitationView
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof IssuedInvitationView
     */
    token: string;
}

/**
 * Check if a given object implements the IssuedInvitationView interface.
 */
export function instanceOfIssuedInvitationView(value: object): value is IssuedInvitationView {
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('expiresAt' in value) || value['expiresAt'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('token' in value) || value['token'] === undefined) return false;
    return true;
}

export function IssuedInvitationViewFromJSON(json: any): IssuedInvitationView {
    return IssuedInvitationViewFromJSONTyped(json, false);
}

export function IssuedInvitationViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): IssuedInvitationView {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': (new Date(json['createdAt'])),
        'expiresAt': (new Date(json['expiresAt'])),
        'id': json['id'],
        'token': json['token'],
    };
}

export function IssuedInvitationViewToJSON(json: any): IssuedInvitationView {
    return IssuedInvitationViewToJSONTyped(json, false);
}

export function IssuedInvitationViewToJSONTyped(value?: IssuedInvitationView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'createdAt': value['createdAt'].toISOString(),
        'expiresAt': value['expiresAt'].toISOString(),
        'id': value['id'],
        'token': value['token'],
    };
}